requirements.txt
alembic==1.3.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==5.0.3
dominate==2.4.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.23.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.8
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
